Clean references and relocate code
