--- napari_hub_cli/citations/__init__.py
@@ -1,3 +1,0 @@
-from .citation import create_cff_citation
-
-__all__ = ["create_cff_citation"]
