--- napari_hub_cli/resources/metadata_usage.csv
@@ -1,17 +1,0 @@
-Field,Filterable,Sortable,Searched
-Authors,False,False,True
-Bug Tracker,False,False,False
-Description,False,False,True
-Development Status,False,False,False
-Documentation,False,False,False
-License,True,False,False
-Name,False,True,True
-Operating System,True,False,False
-Project Site,False,False,False
-Python Version,True,False,False
-Requirements,False,False,False
-Source Code,False,False,False
-Summary,False,False,True
-Twitter,False,False,False
-User Support,False,False,False
-Version,False,False,False
