+++ tests/test_remote_analysis.py
@@ -10,7 +10,7 @@ from napari_hub_cli.checklist.analysis import (
     display_remote_analysis,
     write_csv,
 )
-from napari_hub_cli.checklist.metadata_checklist import (
+from napari_hub_cli.checklist.metadata import (
     AnalysisStatus,
     PluginAnalysisResult,
     analyse_local_plugin,
